Act011_Main_Presenter_Impl  - Modificado metodo deleteFormLocal(GE_Custom_Form_Local) para somente deletar o formLocal e não existir o formData. Caso existe o formData, não apaga e registra como exception

+++ app/src/main/java/com/namoadigital/prj001/ui/act011/Act011_Main_Presenter_Impl.java
@@ -3,40 +3,14 @@ package com.namoadigital.prj001.ui.act011;
 import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
-
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
 import com.namoa_digital.namoa_library.util.HMAux;
 import com.namoa_digital.namoa_library.util.ToolBox;
-import com.namoadigital.prj001.dao.EV_Module_Res_Txt_TransDao;
-import com.namoadigital.prj001.dao.GE_Custom_FormDao;
-import com.namoadigital.prj001.dao.GE_Custom_Form_BlobDao;
-import com.namoadigital.prj001.dao.GE_Custom_Form_Blob_LocalDao;
-import com.namoadigital.prj001.dao.GE_Custom_Form_DataDao;
-import com.namoadigital.prj001.dao.GE_Custom_Form_Data_FieldDao;
-import com.namoadigital.prj001.dao.GE_Custom_Form_FieldDao;
-import com.namoadigital.prj001.dao.GE_Custom_Form_Field_LocalDao;
-import com.namoadigital.prj001.dao.GE_Custom_Form_LocalDao;
-import com.namoadigital.prj001.dao.GE_FileDao;
-import com.namoadigital.prj001.dao.MD_Product_SerialDao;
-import com.namoadigital.prj001.model.DaoObjReturn;
-import com.namoadigital.prj001.model.GE_Custom_Form;
-import com.namoadigital.prj001.model.GE_Custom_Form_Data;
-import com.namoadigital.prj001.model.GE_Custom_Form_Local;
-import com.namoadigital.prj001.model.GE_File;
-import com.namoadigital.prj001.model.MD_Product_Serial;
+import com.namoadigital.prj001.dao.*;
+import com.namoadigital.prj001.model.*;
 import com.namoadigital.prj001.receiver.WBR_Upload_Img;
-import com.namoadigital.prj001.sql.GE_Custom_Form_Blob_Local_Sql_005;
-import com.namoadigital.prj001.sql.GE_Custom_Form_Blob_Sql_001;
-import com.namoadigital.prj001.sql.GE_Custom_Form_Data_Field_MULTI_SqlSpecification;
-import com.namoadigital.prj001.sql.GE_Custom_Form_Data_MULTI_UNIQUE_SqlSpecification;
-import com.namoadigital.prj001.sql.GE_Custom_Form_Fields_Local_Sql_001;
-import com.namoadigital.prj001.sql.GE_Custom_Form_Local_Sql_002;
-import com.namoadigital.prj001.sql.GE_Custom_Form_Local_Sql_003;
-import com.namoadigital.prj001.sql.GE_Custom_Form_Local_Sql_004;
-import com.namoadigital.prj001.sql.GE_Custom_Form_Local_Sql_005;
-import com.namoadigital.prj001.sql.GE_Custom_Form_Sql_001_TT;
-import com.namoadigital.prj001.sql.MD_Product_Serial_Sql_002;
-import com.namoadigital.prj001.sql.MD_Product_Serial_Sql_016;
-import com.namoadigital.prj001.sql.Sql_Act011_002;
+import com.namoadigital.prj001.sql.*;
 import com.namoadigital.prj001.util.Constant;
 import com.namoadigital.prj001.util.ToolBox_Con;
 import com.namoadigital.prj001.util.ToolBox_Inf;
@@ -493,17 +467,64 @@ public class Act011_Main_Presenter_Impl implements Act011_Main_Presenter {
         mView.callAct005(context);
     }
 
+    /**
+     * Modificado em 16/04/2019 por luche
+     * Antes de tentar o delete da form_local, verificar se existe o registro
+     * custom_form_data para o mesmo form.
+     * Caso exista, não permitir o dele.
+     *
+     * @param formLocal
+     */
     @Override
     public void deleteFormLocal(GE_Custom_Form_Local formLocal) {
-        custom_form_LocalDao.remove(
-                new GE_Custom_Form_Local_Sql_005(
-                        String.valueOf(formLocal.getCustomer_code()),
-                        String.valueOf(formLocal.getCustom_form_type()),
-                        String.valueOf(formLocal.getCustom_form_code()),
-                        String.valueOf(formLocal.getCustom_form_version()),
-                        String.valueOf(formLocal.getCustom_form_data())
-                ).toSqlQuery()
+        GE_Custom_Form_Data customFormData = custom_form_dataDao.getByString(
+            new GE_Custom_Form_Data_MULTI_UNIQUE_SqlSpecification(
+                String.valueOf(formLocal.getCustomer_code()),
+                String.valueOf(formLocal.getCustom_form_type()),
+                String.valueOf(formLocal.getCustom_form_code()),
+                String.valueOf(formLocal.getCustom_form_version()),
+                String.valueOf(formLocal.getCustom_form_data())
+            ).toSqlQuery()
         );
+        //Se não existe custom_form_data para esse custom_form_local,
+        //permite a deleção.
+        //Se não, registra exception e NÃO DELETA O custom_form_local
+        if(customFormData == null || customFormData.getCustomer_code() <= 0 ) {
+            custom_form_LocalDao.remove(
+                new GE_Custom_Form_Local_Sql_005(
+                    String.valueOf(formLocal.getCustomer_code()),
+                    String.valueOf(formLocal.getCustom_form_type()),
+                    String.valueOf(formLocal.getCustom_form_code()),
+                    String.valueOf(formLocal.getCustom_form_version()),
+                    String.valueOf(formLocal.getCustom_form_data())
+                ).toSqlQuery()
+            );
+        }else{
+            try{
+                Gson gson = new GsonBuilder().serializeNulls().create();
+                String JsonFormLocal = "";
+                String JsonFormData = "";
+                try{
+                    JsonFormLocal = gson.toJson(formLocal);
+                    JsonFormData = gson.toJson(customFormData);
+
+                }catch (Exception e){
+                    e.printStackTrace();
+                }
+                //
+                String msg =  "Erro ao tentar apagar o registro da GE_Custom_form_local"
+                             + " quando existe um registro na GE_Custom_Form_Data.\n "
+                             + "Situação aconteceu quando o app identificou que o form aberto era novo, não havia "
+                             + " sido salva nem via 'clique na foto',(bNew = true) e o obj formData era != null e status pending.\n"
+                             + " Chamado pelo metodo exitAlert() noOnBackPressed.\n"
+                             + " Dados do custom_form_local: \n" + JsonFormLocal+"\n"
+                             + " Dados do custom_form_data: \n" + JsonFormData+"\n";
+
+                throw new Exception(msg);
+            }catch (Exception e){
+                ToolBox_Inf.registerException(getClass().getName(),e);
+            }
+        }
     }
 
     @Override
